Remove (but save to another file) unused test functions. Update some docstrings.
